Improve error message when sf_format_check fails in Format construction
sf_format_check can fail when the given combination file format/encoding
is not supported or does not make sense. This is actually more likely
than a bug of my own in that case. Since audiolab can noe reliably query
available formats and encodings, if the error message is not consistent
- as in a bug, the user can still realize somethine is wrong.